refactor(shell/sidebar): extract webview module + history panel
Move the WEBVIEW MODULE (cache, partitions, auth URL patterns,
getOrCreateWebview/loadWebviewInPanel/hideWebviews/safeSetPanelHTML) into
shell/js/sidebar/webview.js, and the HISTORY PANEL MODULE (loadHistoryPanel,
renderHistoryItems, attachHistoryClickHandlers, loadSyncDevices) into
shell/js/sidebar/panels/history.js.

Callers inside index.js that reached into webviewCache directly now go
through hasWebview()/getWebview() accessors (same pattern as config.js).
COMMUNICATION_IDS moves alongside the webview module since it's the set
of items that open as sidebar webviews.

Behavior unchanged. No bug fixes, no renames, no rewrites — pure
extraction per scope discipline §5.3.

Part of the 12-PR shell/js split (PR 9/12).
